Arreglados bugs en batalla que detenían la ejecución de la misma, relacionados con las habilidades con prioridad. Añadido log al BattleAbilityManager. Efectos de habilidades se muestran ahora con decimales correctos. No se puede mirar hacia abajo si no estás en el suelo. Arreglado progreso de subida de nivel de enemigos por el mapa. Actualizado Assets.psb con nuevas escaleras. Otros cambios menores.

+++ Assets/02. Scripts/Battle/EffectImplementations/BattleEffectBlind.cs
@@ -47,7 +47,7 @@ namespace Laresistance.Battle
 
         public override string GetShortEffectString(int level, EquipmentsContainer equipments)
         {
-            float modifier = GetPower(level, equipments);
+            int modifier = GetPower(level, equipments);
             return string.Format("{0}%", modifier.ToString());
         }
